2D & 3D

+++ src/OOPS/Inheritance/Shapes/ThreeD/Sphere.java
@@ -13,7 +13,7 @@ public class Sphere {
     }
 
     public double Volume() {
-        return (4 / 3) * Math.PI * radius * radius * radius;
+        return ((double) 4 / 3) * Math.PI * radius * radius * radius;
     }
 
     public String toString() {
